负数问题

+++ api/src/main/java/com/juhai/api/controller/OrderController.java
@@ -205,10 +205,10 @@ public class OrderController {
         obj.put("userName", order.getUserName());
         obj.put("projectName", order.getProjectName());
         obj.put("realName", order.getRealName());
-        obj.put("amount", order.getAmount().intValue());
+        obj.put("amount", order.getAmount().longValue());
         obj.put("incomeRate", order.getIncomeRate());
         obj.put("limitTime", order.getLimitTime());
-        obj.put("forecastReturnAmount", order.getForecastReturnAmount().intValue());
+        obj.put("forecastReturnAmount", order.getForecastReturnAmount().longValue());
         obj.put("IdCardNo", user.getIdCard());
         obj.put("orderTime", order.getOrderTime());
         obj.put("forecastReturnTime", order.getForecastReturnTime());
